+++ pydidas/gui/windows/scan_dimension_information_window.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -12,14 +14,13 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with Pydidas. If not, see <http://www.gnu.org/licenses/>.
-
 """
 Module with the ScanDimensionInformationWindow class which explains the scan dimensions.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["ScanDimensionInformationWindow"]
@@ -102,8 +103,8 @@ class ScanDimensionInformationWindow(PydidasWindow):
             layout_kwargs=dict(
                 contentsMargins=(0, 0, 0, 0),
             ),
-            sizePolicy=constants.EXP_FIX_POLICY,
-            alignment=constants.QT_DEFAULT_ALIGNMENT,
+            sizePolicy=constants.POLICY_EXP_FIX,
+            alignment=constants.ALIGN_TOP_LEFT,
             parent_widget=None,
         )
         self.create_any_widget(
@@ -111,7 +112,7 @@ class ScanDimensionInformationWindow(PydidasWindow):
             ScrollArea,
             widget=self._widgets["canvas"],
             fixedWidth=650,
-            sizePolicy=constants.FIX_EXP_POLICY,
+            sizePolicy=constants.POLICY_FIX_EXP,
             stretch=(1, 0),
             layout_kwargs={"alignment": None},
         )