+++ pydidas/widgets/dialogues/acknowledge_box.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -19,18 +21,20 @@ box to hide this dialogue in the future.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["AcknowledgeBox"]
 
 from qtpy import QtCore, QtWidgets
 
-from ...core.utils import get_pydidas_icon
-from ...core.constants import EXP_EXP_POLICY
-from ...core.utils import format_input_to_multiline_str
-from ...core.utils import apply_qt_properties
+from ...core.constants import POLICY_EXP_EXP
+from ...core.utils import (
+    apply_qt_properties,
+    format_input_to_multiline_str,
+    get_pydidas_icon,
+)
 from ..factory import CreateWidgetsMixIn
 from ..scroll_area import ScrollArea
 
@@ -67,7 +71,7 @@ class AcknowledgeBox(QtWidgets.QDialog, CreateWidgetsMixIn):
         apply_qt_properties(
             self._widgets["label"],
             textInteractionFlags=QtCore.Qt.TextSelectableByMouse,
-            sizePolicy=EXP_EXP_POLICY,
+            sizePolicy=POLICY_EXP_EXP,
             indent=8,
             fixedWidth=500,
         )