+++ pydidas/widgets/dialogues/user_config_error_message_box.py
@@ -30,7 +30,7 @@ import os
 from qtpy import QtCore, QtWidgets, QtSvg
 
 from ...core.utils import get_pydidas_error_icon_w_bg, get_pydidas_icon_path
-from ...core.constants import EXP_EXP_POLICY
+from ...core.constants import POLICY_EXP_EXP
 from ...core.utils import format_input_to_multiline_str, apply_qt_properties
 from ..factory import CreateWidgetsMixIn
 from ..scroll_area import ScrollArea
@@ -67,7 +67,7 @@ class UserConfigErrorMessageBox(QtWidgets.QDialog, CreateWidgetsMixIn):
         apply_qt_properties(
             self._widgets["label"],
             textInteractionFlags=QtCore.Qt.TextSelectableByMouse,
-            sizePolicy=EXP_EXP_POLICY,
+            sizePolicy=POLICY_EXP_EXP,
             indent=8,
             fixedWidth=500,
         )