+++ pydidas/widgets/scroll_area.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -12,22 +14,22 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with Pydidas. If not, see <http://www.gnu.org/licenses/>.
-
 """
-Module with the ScrollAreaclass , a QScrollArea implementation with convenience
+Module with the ScrollAreaclass, a QScrollArea implementation with convenience
 calling arguments for simplified formatting.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["ScrollArea"]
 
+
 from qtpy import QtWidgets
 
-from ..core.constants import EXP_EXP_POLICY
+from ..core.constants import POLICY_EXP_EXP
 from ..core.utils import apply_qt_properties
 
 
@@ -51,7 +53,7 @@ class ScrollArea(QtWidgets.QScrollArea):
         super().__init__(parent)
         kwargs["widgetResizable"] = True
         kwargs["autoFillBackground"] = True
-        kwargs["sizePolicy"] = EXP_EXP_POLICY
+        kwargs["sizePolicy"] = POLICY_EXP_EXP
         kwargs["frameShape"] = QtWidgets.QFrame.NoFrame
         apply_qt_properties(self, **kwargs)